+++ webpack.config.js
@@ -6,6 +6,9 @@ module.exports = {
   entry: {
     bundle: path.resolve(__dirname, './src/index.js'),
   },
+  devServer: {
+    static: './dist',
+  },
   output: {
     path: path.resolve(__dirname, 'dist'),
     filename: '[name].js',
